# level 5 assignment/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/rekening": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Rekening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rekening"
                ],
                "summary": "Put Rekening Data",
                "parameters": [
                    {
                        "description": "Rekening",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateRekeningRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add new Rekening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rekening"
                ],
                "summary": "Post Rekening Data",
                "parameters": [
                    {
                        "description": "Rekening",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.SaveRekeningRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/rekening/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get All Rekening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rekening"
                ],
                "summary": "Get All Rekening Data",
                "responses": {}
            }
        },
        "/rekening/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Rekening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rekening"
                ],
                "summary": "Get Rekening Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Rekening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rekening"
                ],
                "summary": "Delete Rekening Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rekening ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/showFile/{filename}": {
            "get": {
                "description": "Show a file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Show a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filename",
                        "name": "filename",
                        "in": "path"
                    }
                ],
                "responses": {}
            }
        },
        "/training": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Training",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training"
                ],
                "summary": "Put Training Data",
                "parameters": [
                    {
                        "description": "Training",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateTrainingRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add new Training",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training"
                ],
                "summary": "Post Training Data",
                "parameters": [
                    {
                        "description": "Training",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.SaveTrainingRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/training/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get All Training",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training"
                ],
                "summary": "Get All Training Data",
                "responses": {}
            }
        },
        "/training/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Training",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training"
                ],
                "summary": "Get Training Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Training",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training"
                ],
                "summary": "Delete Training Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Training ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/upload": {
            "post": {
                "description": "Uploads a file and saves it",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Uploads a file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "dtos.RekeningKaryawan": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "dtos.SaveRekeningKaryawan": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dtos.SaveRekeningRequest": {
            "type": "object",
            "properties": {
                "jenis": {
                    "type": "string"
                },
                "karyawan": {
                    "$ref": "#/definitions/dtos.SaveRekeningKaryawan"
                },
                "nama": {
                    "type": "string"
                },
                "rekening": {
                    "type": "string"
                }
            }
        },
        "dtos.SaveTrainingRequest": {
            "type": "object",
            "properties": {
                "pengajar": {
                    "type": "string"
                },
                "tema": {
                    "type": "string"
                }
            }
        },
...
        "dtos.UpdateRekeningRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jenis": {
                    "type": "string"
                },
                "karyawan": {
                    "$ref": "#/definitions/dtos.RekeningKaryawan"
                },
                "nama": {
                    "type": "string"
                },
                "rekening": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateTrainingRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "pengajar": {
                    "type": "string"
                },
                "tema": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}